swal + grafica

+++ src/app/pages/party/party.component.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { Location } from '@angular/common';
 import { PartyService } from 'src/app/service/party.service';
-
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-party',
   templateUrl: './party.component.html',
@@ -11,11 +11,13 @@ import { PartyService } from 'src/app/service/party.service';
 export class PartyComponent {
 
   partys : any
+  close : boolean = true
 
   constructor(
     private router: Router,
     private location : Location,
-    private fireBaseSrv : PartyService
+    private fireBaseSrv : PartyService,
+    
   ){}
 
 ngOnInit (): void {
@@ -45,14 +47,36 @@ this.partys = Object.keys(data).map((key)=>{
 
 
 deleteParty(id:any){
-this.fireBaseSrv.deleteParty(id).subscribe(data=>{
-  console.log(id);
-  
-  console.log(data);
-  if(data==null){
-this.getParty()
-  }
-})
+  Swal.fire({
+    title: 'Sei sicuro di voler eliminare?',
+    showDenyButton: true,
+    // showCancelButton: true,
+    confirmButtonText: 'Conferma',
+    denyButtonText: `Indietro`,
+  }).then((result) => {
+    /* Read more about isConfirmed, isDenied below */
+    if (result.isConfirmed) {
+      this.fireBaseSrv.deleteParty(id).subscribe(data=>{
+        console.log(id);
+        
+        console.log(data);
+        if (data==null){
+          Swal.fire('Eliminato!', '', 'success')
+          this.getParty()
+        }
+          else{
+            Swal.fire({
+              icon: 'error',
+              title: 'Oops...',
+              text: 'Qualcosa è andato storto, elemento non eliminato',
+              // footer: '<a href="">Why do I have this issue?</a>'
+            })
+          }
+      })
+      
+    } 
+  })
+
 
 }
 
@@ -61,10 +85,21 @@ goToAdd(){
   this.router.navigate(['party/add-party'])
 }
 
+goToEdit(id:any){
+this.router.navigate(['party/edit-party/' + id])
+}
 
 goToBack() {
   this.location.back();
 }
+
+
+closeDrop(){
+ this.close = !this.close
+ console.log(this.close);
+ 
+}
+
 
 
 }
